Revert split of TestLldbGdbServer.py
